home.php: cleaned up loading of libraries and models. now use a single line to build tabs instead of two lines.
navtab.php: modified make_tabs() function to accept an optional argument in which the tabset is passed allowing the tabs to be returned in a single function call instead of a minimum of two.

manage2.php: initial check-in of a page exactly like manage.php which is used by the managecourses.php controller.