The get_enum_vals method now gets the current database from the db object.
Changed the get_enum_vals method in dbmetadata.php to determine which
database was being used by a particular instance of the db from the
db->database value.

Changed the calls to get_enum_vals in materials.php to use the 2
argument version of the method instead of the older 3 argument
version.