Added fix for select and delete buttons shifting table to the right.
The select, unselect and delete items buttons were throwing the table to the
right of the page. The table was sitting next to the buttons not below it.
This behavior was not seen in Firefox 3 RC but was seen with Safari 3.1.1
and with Firefox 2. A series of <br /> tags were inserted below the css
div that holds the buttons. A cleaner fix is needed.