Fixed bugs in the download materials code and improved file naming.
The download materials code had a misplaced call to the function that
triggered the zip file download which was causing it to not trigger
the download for less than 10 materials.

The download function was refactored so there is a LOT less replicated
code.

The file naming was fixed so there should be little if any file/folder
name collisions.