Fixed date display issue in the materials table reported in OERDEV-27.
Moved mysql time comparison code out of the materials.php controller to
a new function "calc_later_date" in ocw_utils and made adjustments in code
that depended on it.

The "materials" and  "get_material_paths" functions in the material.php model now
use "calc_later_date" and return the later of created or modified timestamps in
human readable format as part of their result sets.