The progress bar and label on the dscribe1 home link to course materials
The progress bar for each course on the dscribe1 home page now links to
the materials table for that course. Users can simply click on the
image or on the label and be taken to the course materials page.
The homedscribe1.php view was modified to make this change.

Added comments where there were none in the home.php controller.

Commented the "get_co_count" and related functions in the material.php
model.

Commented the "get_courses_simple" and "get_dscribe2_rel" functions in
the ocw_user.php model.