Fixed trailing comma and js method calls
The IE6 js interepreter did not like trailing commas at the end of
method calls and function calls. It also balked at trailing commas
after the last function/method in an class.

Methods calls on js objects require that the js object be noted as a
variable and surrounded by braces. e.g.
this.parentNode.removeClass(rmvclass) changed to
$(this.parentNode).removeClass(rmvclass)