More changes for OERDEV-120
Move the code that destroys a "stale" session into the check function.
From there, redirect to the originally requested location.  If it turns
out that the "new" user is not allowed to go where they have requested,
the denyAccess code will now do the right thing and redirect them to
their home location and flash the access denied message.

Also, make sure that things work in a local test environment where
"remote_user" is set in the config.php file.