Clean up progress bar code and comments.
Commented out function and code blocks in the progress bar
library and in the home.php function where the library was
called were cleaned up. The template call was also modified
to get around a bug in the buggy menu building library and
uses the older, more reliable code in the header template.