OERDEV-136
This adds admin interfaces to create schools, subjects, curriculum, and courses.
It modifies the existing code that edits course information.  Adding and editing
a course now share a large chunk (or three small chunks) of code.

There are still a couple of issues that Piet would like changed:
 - Change "edit user" and "edit profile" such that
   they appear as a single dialog as is done when
	 adding a user
 - Add the ability to add curriculum and subjects when
   creating a school (or when creating a course)

Meanwhile, this code can be scrutinized and exercised.