Set default CO view to all when no new objects
Jira bug 164.

If there are any content objects without assigned status, then
the CO view for a particular material shows those "new" objects
in the initial view when clicking on the link for that material.

However, if all objects for a material have assigned status which
means there are no "new" objects, then all objects are visible in
the initial view (All) when clicking on the link for that material.