CO display was being forced to "New"
The fix for OERDEV-164 was incorrect. The CO display was being forced to
display the "New" view regardless of what was selected. This fix changes
that so that if no view is specified and no new COs are present, it
displays the "All" view and when a view is specified it displays that
view.