OERDEV-178
Allow a custom prompt in the confirmation dialogs when the admin
is deleting users, curriculums, subjects, schools, or courses.

Changes the 'confirm' Javascript class to look for a 'customprompt'
attribute.  If none was specified, fall back to the default prompt,
"Are you sure?".

Change admin uses of the confirm class to supply custom prompts.