OERDEV-192
This is a quick & dirty fix.  The "real" fix is to create the iframe
with a size relative to the original window size.

This changes the size of the edit iframe to be short enough to
fit in an 800x600 browser window.

It also adds code to make sure that the requisite additional
fields become visible when changing the recommended action.

(It also changes the way the window is scrolled when the
"ask question" radio button is changed to yes.)