OERDEV-192
The addition of the vertical scroll bar causes the window to be
too narrow, causing a horizontal scroll bar as well.  Add extra
width to account for this.

The edit course, add course frame width is already set to 800,
so there is no avoiding this issue there without re-designing
the layout to fit within a narrower space.