OERDEV-192
This is a better fix for the problem of iframes being too big for the parent window.

If the requested size is too big to fit within the parent window, then
resize it to an appropriate size.

Committing this as an independent revision from the restoration of all the iframe
sizes.