OERDEV-177
When editing a material's information, if the material has Content Objects,
disable the "Embedded COs?" radio button.  In this case, it should be marked
as "yes" and can only be changed to "no" if all the COs for the material have
been deleted.

If the material has no COs, leave the button enabled to go between "yes" and "no".