Remove white titlebar on smoothbox panels
OERDEV-195. This is part one of a two part fix.
The form header test was repeated at the top of the smoothboxes.
This text was inconsistent in some cases and redundant in
others. Since this deals with a javascript and the rest of the
fix is edits to forms, I decided to make two separate commits.