Changed wording of "No copyright" CO status/action
OERDEV-205. The OER Policy group determined that instances or
"Retain: No copyright" and "No copyright" should be changed to
"Retain: Copyright Analysis" and "Copyright Analysis".

"Re-Create" was changed to "Create".
Along with the code changes, the update_action_items_1.sql script
was added to the the dbscripts folder in the SVN tree to facilitate
database schema changes and record updates required to keep the DB
consistent with these wording changes.