Enable download of large material collections
OERDEV-169. The codeigniter zip library and download library are memory
inefficient. The archive creation functionality was abstracted out and
zip archives are now created using the php ZipArchive class, which seems
to use memory more efficiently. Downloads are now doing buffered file
reads instead of reading the whole file into a string and then sending
it to the browser.