OERDEV-106
Note that the code changes allow larger images to be captured
by not always keeping a copy of the last image around.  They
also do a better job of catching and handling the out-of-memory
condition.  This makes it possible to recover from the condition
w/o requiring a restart of the browser.

Changes to the Snapper java code:
- Remove the call to detectChange() which is only
  useful when doing auto-capture
- Add calls to explicitly release memory when no
  longer needed to make it eligible for Garbage Collection
- Catch the out of memory exception and return an
  indication of an error to the javascript code

Changes to the javascript code:
- Handle the out of memory error from getBase64Jpeg()
- Handle an error sending image to the server

Add file snapper-java.diff with the changes made to the
java code.

Add file java-deployment-properties.diff showing how
to change (on a Mac) the maximum memory that the applet
can use.  This allows large images (larger than
approximately 2.4MB) to be captured.