Cleaned up download of single material.
The URL to the material file previously did not include the file extension
causing some configs of apache to not download the file and others to
mysteriously tack on the file extension and serve the first file with
an extension that matches on the list of mime types provided to the web
server. This has been changed so the the full name of the file is used
which doesn't require that the extension be appended. Likely more secure.