Confirm existence of directory before scanning.
Introduced checks to see if the directory exists at the specified path
before calling the scandir() function on that path. scandir()
throws errors when run on non-existent directories which disrupts the
material download process because errors are printed to screen.

We confirm that the directory exists before calling scandir() and
proceed only if the directory does exist.