Log more data during archive creation.
The archive creation process, particularly within the _build_zip function
now captures more data during the archive creation process. A bunch of
code is also present, but commented that will send information to the
apache error_log. Uncomment the code to get fairly comprehensive info
in the apache_error log that traces the archive creation process.

Also moved the call to ZipArchive::close() so it is only called if the
ZipArchive::open() call has worked.