Added code to help debug material download
Material download has been failing due to a bug in PHP. This patch
adds a $DEBUG variable to the _build_zip function in the
oer_create_archive.php library. When set to "TRUE" a bunch of
DEBUG information is sent during archive creation to the apache
error log.

The file download code also checks to see if the selected file or
archive containing selected files exists before continuing and exits
if the file/archive doesn't exist.