Enhanced logging in oer_create_archive library.
Changed logs around so that some output is sent to the CI log. This may
appear to be a mess, but certain errors are simply not printed to the CI
log and we have to resort to the apache log. In this case, if the zip
archive isn't opened, we print to the apache log and abort the script.