Fixed multiple material download.
Multiple material downloads weren't working after the upgrade to
PHP 5.2.8 because the call to the ZipArchive::open method which
in our case was $zip->open was using the ZIPARCHIVE::OVERWRITE
flag. We needed to pass it the ZIPARCHIVE::CREATE flag. Thanks
to Kevin, we are passing it both and the call is now:
$zip->open($filename, ZIPARCHIVE::CREATE | ZIPARCHIVE::OVERWRITE)
which works.