Improved course addition form validation
The course add form was not correctly validating form length because the
default "-- select --" form field had a value of "-1" which was causing
the validation to consider it valid data and the course was as a result being
added.

We hacked in the course_length_check() function which is called when
validating and checks for the "-1" value, sets it to NULL and changes
fails validation.

When getting the cid value for a newly added function, we match on more
fields for the newly added record, which reduces the chance of matching
incorrectly. We were previously only matching on course title which could
cause problems when two courses had the same titles.

Additional stuff related to course addition has been cleared up.