=============================================== Add initial support for Document Decomposition ===============================================
1) Add a new function, get_extension_by_mime(), to complement
the get_mime_by_extension() function.  Use it to go back to an
extension from a given mimetype.

2) Move function manually_add_materials() from models/material.php
into controllers/materials.php so that the decomposition
stuff can be called from an controller rather than a model.

3) Add code to attempt material decomposition when adding new
materials.

4) Add new libraries for decomposition.  For Microsoft documents
which are handled by the apache_poi libraries (.ppt and .doc),
use poi to extract images.  For other documents, call the
python hachoir code to attempt to extract images.

Note that new options are added to the config/layout.template!!