Create thumbnails when adding new images or displaying existing images
When new a new image is added, create a thumbnail.  When building a page
of images, if thumbnails don't exist, create them.  (If unable to create
a thumbnail, a symlink is created so that it isn't continually retried.)

Add a function to remove previously existing replacement files.  Call it
in two places.

Update the replacement_exists() function to avoid building paths which
are never used.  Use a loop over supported file extensions rather than
a list of repeated names.

Signed-off-by: Kevin Coffman <kwc@citi.umich.edu>