Changed dscribe2 admin page to use morpthabs.
The "admin/users/edit/dscribe2/<userid>.html" page
e.g. "admin/users/edit/dscribe2/4.html"
was modified to use morphtabs in place of mootabs. The new page
was not displaying the full width of the table, instead it was
restricting width to the value dictated by morpthabs as set in the
width and height arguments defined in the setuptabs function in
ocwui.js I set those to blank which fixed the width issue.

A background color bug still remains. The background of the page
should be white instead it is green. The active tab is also green and
should be grey instead. Instead the inactive tab is grey.