Switched back to morphtabs.js on dscribe 2 admin.
The jquery tab plugin changes the look and feel too much. Integrating will be
a pain as well. So, we switch back to the morphtabs.js plugin. This is an
intermediate step to facilitate the writing of our own tabbed interface plugin.

Also added the oercatabs.js file which is the start of a replacement for
morphtabs.js which is not currently usable. It won't require that we pass
explicit size values for the tabbed interface.