Fix oercatabs.js bugs with IE.
The event 'target' method wasn't available because it does not propogate in IE
due to performance reasons. It required use of the '$' selector around the
'item.target' call which then made the 'getParent' method available.

There is still a bug with IE 6 where clicking on a part of the tab other than
the text causes an error. But it's IE 6 for the LOVE OF GOD. Will fix when we
have a bunch of free time.