OERDEV-219 / OERDEV-227
Display dscribes associated with a course in the course listing

Replace the "Primary Instructor" column with a column listing the
dScribes associated with the course.

Modify the get_course_instructor_by_id() function to take
an additional parameter of the type of users to return.
Rename the function to get_course_users_by_cid().

Call it for dscribes as well as instructors.