OERDEV-229
Remove the actions, "Permission", "Fair Use", and "Commission" from the
list of actions available to be chosen by dScribes as recommended or
final actions.

Since there are existing entries in the database using these actions, they
cannot be completely removed from the database.  They should simply not be
available for future selection.

Update the RAD decision tree to no longer recommend an action of "Permission".
Instead, instruct the describe to obtain the necessary permissions and select
"Retain: Permission".