OERDEV-233 Some Ask Form metadata is not being saved when changed
It turns out that the "Content-Type" and all copyright information was not
being saved.

Change the Content-Type (subtype) do use class do_ask_object_update rather
than do_object_update since "oid" is not available.  (It is encoded in the
name.)

Modify the InPlaceEditor JavaScript function to take a parameter specifying
whether to use 'get' or 'post' to change the value.  The update_object
controller function expects 'get', while the update_object_copyright
controller function expects 'post'.  The 'post' method expects that the field
name to be modified is included in the parameters as well.

Change all the uses of InPlaceEditor to specify 'get' or 'post', and modify
the copyright callers to use the right parameters.