Fix 'next'/'previous' CO counting error.
OERDEV-240.
The incorrect regexp match result was being used for the array key string to
select the COs for the specified $filter argument.

The $m regexp results array has two results. The first result (element 0)
matches the a key in the $s['objects'] array while the second result
(element 1) did not match. The code was changed to use the first match
result (element 0).