Remove unneeded horizontal scrollbar in CO edit.
The CO details lightbox was displaying unneeded horizontal scroll bars because
the column widths used were too wide. The column widths were changed from
24 columns (based on the blueprint css library column width where 24 columns
is a width of 950ps) to 20 columns which is within the horizontal width of the
lightbox.

This is a fix to the fairly sparse spec for OERDEV-236.