Further changes to object_info panel.
The Ask Instructor and Ask dScribe2 question radio buttons reveal additional
links when the "Yes" buttons are selected. The revised version only reveals
a link to the Ask form instead of the 3 previous links.

The dynamic form that appears depending on choice of recommended actions
was relocated to the position below the Recommended Action pull down menu.