Use the "Nested Accordion" code from: http://blog.medianotions.de/en/articles/2008/mootools-nested-accordion to overcome the issue of nested FX.slides not getting the size calculation correct.
I modified the original JavaScript code to use a for() loop rather than a
while() loop to fix an issue where the nested accordions were not being
properly recognized.

Note that the accordion code doesn't like embedded divs in the accordioned
stuff.

I also changed the ":" to a "?" after the "Ask instructor if replacement is
suitable" text.