Change question ask functionality in CO details.
OERDEV-239.

The 'Ask dScribe2 a general question about the CO' 'Yes' radio button now
displays a text entry box and a submit button.

The 'Ask a general question' functionality has been removed and it is no longer
possible to ask the Instructor a question. This was removed at Piet's request.