OERDEV-250 PHP error in askform for dscribe2 question about replacement
OERDEV-239 (and OERDEV-252) changed the CO edit forms.  The dscribe is
no longer allowed to ask the instructor general questions about the
replacement object.  Instead, they should ask the dscribe2.  While
there was functionality to ask the dscribe2 before, it wasn't the default and
was apparently rarely used.  These changes fix the dscribe2 askform view for
replacement object questions.

It also changes the workflow slightly because the old workflow was broken
(IMHO).  An instructor or dscribe2 could select the "Send to dscribe" button
without ever answering a question.  Yet, it would be marked as "done".