A misunderstanding of the specs...
Move the "Ask dScribe2 a general question" above the
"Ask instructor if replacement is suitable".

Don't show the dScribe2 the "Is this replacement
suitable" question.

Also fixes the email notification when the dscribe2 answers a replacement
question.  It was previously 'hardcoded' as coming from the instructor.  It
now checks the role and tries to do the right thing.