Display both images in "Addressed Items" for replacement objects
OERDEV-246
OERDEV-250

The dScribe2 askform view of "Addressed Items" should display
the original object in the first column, and the replacement
in the second column.

There was a PHP error in the navigation arrow generation.  Changed
the prev_next() function in a couple of ways:
1) When selecting anything but "All" in the Response Type dropdown
   in the dscribe2 askform view, the filter became something like,
   "general:general".  Add a check for this case.
2) There is a mixture of original objects and replacement objects
   in the array when viewing all addressed items.  When looking for
	 previous and next, we must check the type of object for each case.

There was missing functionality for the non-edit displays.  If an
object is 'in progress' there was nothing displayed.  Added text to
indicate that the object is under review.

While testing, discovered how to fix the problem described in OERDEV-246
where the status "flag" was appearing in the wrong place.  Needed to
specify the 'offset' parameter in the call to create_co_img() for
'in progress' objects.