Use oer_faceted_search library for facet options.
The oer_faceted_search library provides a function get_facet_options which is
called in the controller that generates the page containing the faceted
search. These modifications have allowed the removal of model (DB) calls
in the views and reduce business logic in the view code.

This commit does not restore faceted search functionality at the course
level.