Escape field names and values in AJAX calls.
OERDEV-258

The ampersand characted and any text after it was being truncated in the
object location field in the CO details iframe because the field names
and values weren't being escaped before the POST values were defined.

This patch escapes both the field name and the field value so that the
ampersand may be used in either one and will be passed on without
truncation.