Fix problem where not all RCOs downloaded.
OERDEV-253.

The case of the extension of the CO file name is not the same in the DB and
on the filesystem. Previously we would only match against the value in the DB.
We now match against the value in the DB and both uppercase and lowercase
variants of the extension which increases our chances of finding all files.