Prevent status key from overlapping with CO view.
OERDEV-261.

The status color key restored in SVN r61798 overlapped with the CO view if a
filter was selected. These changes prevent them from overlapping.

WARNING, there are more hard coded style values in the _edit_material_cos.php
file as a result of this commit. This is bad practice, but we are simply adding
to the bad, not starting the bad.