Make CO status counting consistent throughout
Fixes:
OERDEV-259: The "No Action Assigned" count overlaps with "Cleared" and "In Progress"
OERDEV-260: Define consistent criteria for CO status indication


Change several places to use the agreed-upon criteria for determining the
status of a CO.

The Criteria:
=========
"cleared" -- An object is considered "cleared" when a final action
(action_taken) has been selected AND the object is marked as "cleared for
publishing" (done).

"in progress" An object is considered "in progress" when it is not "cleared"
AND any one of four things has been done with the object:
  1) A question has been asked of the instructor (ask == yes)
	2) A question has been asked of the dScribe2 (ask_dscribe2 == yes)
	3) A Recommended Action (action_type) has been selected (is not NULL)
	4) A Final Action (action_taken) has been selected (is not NULL)


Add code to set ask_status or ask_dscribe2_status in the object when
a question about that object is actually asked.

A previous commit enforces that an object cannot be marked as cleared
unless a Final Action has been selected.