OERDEV-263 Fix prev/next arrow navigation
The old prev_next calculations were dependent on a static array
of objects and their status.  With faceted search, a subset of
objects can be selected, and the static arrays no longer worked.

Add a hidden field to the content object display pages with the
object ids of the subset of objects being displayed.  Use this
array to calculate the prev/next arrows.

The askforms also needed to be adjusted to include the hidden
field since the same CO edit page is used in both cases.  Of
course, the askform code is different for questions asked of
dscribe2 vs. instructor, so that complicated the matter.

Ideally, the arrow generation should be done all in javascript
rather than the Ajax call that is currently used.  But enough
is enough...