OERDEV-264: Don't use ask_form_info data for object_stats!
The implementation of object_stats uses the output of ask_form_info().
This fails to properly categorize objects by recommended action since the
ask_form_info depends on whether the object_claims have been handled or not.

For example, an object with a recommended action of "Retain: Copyright
Analysis" which has all the questions answered will be put in the Addressed
Items ("aitems") bin by ask_form_info().  When displaying the normal object
edit page (faceted search), that object should be in the "retain:ca" bin.

The mistake is that we are simply copying the "askinfo" data, which should not
be used when dispaying data for the object edit page...
Simply count the 'Fairuse', 'Permission', 'Commission', and 'Retain: Copyright
Analysis" items like we do the other recommended actions.

While in here, fix the previously noted error where all objects were being
added to the "uncleared" bucket.  Instead, only add them there if they are
not cleared!!  (A radical concept...)