Clean up get_courses() function(s)
The remaining callers of model function course->get_courses()
do not need all the material statistics.  This is now collected
in new_get_courses() for those that need it.  Remove this from
the get_courses() function since it was taking upwards of 30
seconds to get a simple list of all the courses for an admin
user.

Remove the ocw_user->get_courses() function since it is not
used and just causes confusion.  Also remove old commented
out invocations of this function.  (They should now be using
course->get_courses() or course->new_get_courses().

The only use of model function course->faceted_search_get_courses()
is the controller function courses->faceted_search_courses().
It is not clear that this controller is ever used.  If not,
we could eliminate even more cruft!