Add get_created_objects.php script.
This script is quite limited. It queries the DB for any cleared objects
that have an action_taken value of "Create" and grabs all the associated
tags for those images.

It then creates a zip archive of those messages
and also creates a metadata file containing the original file name and
the tags for that file.

The metadata file is in a delimited format where the delimiter is ';;'
(two semicolons).