Bypass issue with convert for wmf files on RHEL 5.4 / CentOS
Running 'convert' (ImageMagick program) on emf (or certain
emf) files causes a call to mmap with a huge size.  This
sends the system into a frenzy trying to map that much
virtual memory.  The size in the mmap call is 4446617600
(0x1090a0000).

As a quick bypass, just ignore wmf image files.
(Just don't try to convert them for now...)

Signed-off-by: Kevin Coffman <kwc@umich.edu>