Initial commit delete_extra_records.sql
This file has a set of canned queries that facilitate the deletion of
all records that aren't relevant to the specified course IDs. Not sure
if it works perfectly yet, but appears to mostly do the job.

I've also squeezed in some whitspace cleanups and removals of hard
tabs in oer_decompose.php