Fix OpenOffice invocation for apache user on Linux
The apache user has no shell, and is not allowed to create
files or directories in its home directory.  When invoking
the OO decomp program, set HOME to /tmp so that necessary
configuration information can be created and referenced by
the apache user.

Also note that the registration wizard must be disabled.
Otherwise, the OpenOffice instance is never ready to accept
connections, and the UNO code is unable to connect to the
soffice program that is started.  See:
http://wiki.services.openoffice.org/wiki/Documentation/Administration_Guide/Deactivating_Registration_Wizard
for more information on how to disable this PITA!