Add CitationManipulate class.  Add boilerplate option
From Ali Asad Lotia:
Add new CitationManipulate class to handle parsing of
a citation string.  Remove local parsing methods from
DecompImpress.

Add command-line and json option to specify the location
of the boilerplate slides to be added to the front of
a PowerPoint slide when saving.  Ali supplied a good
start at this.