Add co image remove on recomp.
CO images with a "Remove and Annotate" final action or recommended
action in cases where there is no final action are replaced with an
image that states that the image has been removed. This informational
image replacement is accompanied by a citation that must be modified by
the dScribe.

The material model also contains setters to specify a different removed
image placeholder from the system default. It also contains a setter for
the accompanying citation.

Downloaded files are now always delivered in zip form even if a single
file is selected for download. This change was a quick way to make sure
that the recomped version was delivered in cases where a single file was
selected for download.

Added a standard image that is put in place of removed images.

Added configuration directives to
tool/system/application/config/layout.template for configuring the path
to the "removed image" replacement file and also to specify the
citation text for removed images.