Add citation instruction tooltip in object edit.
The "Source Information" dialog box in both object and replacement edit
screens now has a tooltip that brings up instructions on citation
formatting.

Making stuff align required some very dodgy css. This css is inline. The
current template code is so hideous that attempting to do things the
right way results in everything getting ugly. So for now, this is an
example in how NOT to do css or tooltips for that matter.