Make the shape/image number zero-based
The old decomp code had page numbers starting at one and image
numbers starting at zero.  Remove the code in the OpenOffice
decomp code that offset the image number to make them one-based.
Update the recomp code to expect zero-based image numbers.

This allows the recomp code to work with materials that were
decomposed with the older decomp code and the new OpenOffice
decomp code.

Remove the adjustment in _get_rec_place_info() that was there
to make the recomp code work for the older material.