Fix off-by-one issue with citations, fix image removal issue
After changing the off-by-one issue with the shape numbers,
the citations had an issue.  This resolves that.

After adding code to allow replacement of non-GraphicObjectShape
shapes, an issue arose when several replacements or citations
occur on the same page.  To replace the non-GraphicObjectShape
we remove the original shape and add a new GraphicObjectShape.
The removal of the original shape causes all the following
shapes on the page to get shifted up.  This causes problems
when one of those images needs to be cited or replaced since
its index has changed.  To fix this, we delay the removals
until all the other operations on the document are complete.