Config changes to allow running outside webroot.
The layout.template file contained several lines that had
$_SERVER['DOCUMENT_ROOT'].'/@OERTOP@'
where @OERTOP@ was replaced with whatever subdirectory of the webroot
in which OERca was installed. However, when OERca was installed outside
the webroot several things break and CO images etc. no longer show up.

This change requires that the user specify the absolute path to the
OERca install directory near the top of the layout.php file and things
should work instead of having to replace all occurrences of @OERTOP@.

Note, it should be the directory where the OERca webapp is installed.
Not the top level of the SVN/Git repository. The name for this folder in
the SVN repository is "tool".