OERDEV-280: Not extracting images extracted by old decomp code
The apache_poi decomposition code extracted certain images as
.pict format files.  The OpenOffice decomp code was extracting
these as .svm (StarView Metafile) format files.  These images
could not be converted using ImageMagick's 'convert' program
because it cannot handle the .svm format.

Change the extraction code to detect if an SVM image is about
to be extracted.  If so, export the image as a .PNG file from
the decomp code.  This also relieves the post-processing code
from having to deal with these images at all.